+++ docs/src/surfaces/actionBar/actionBar.jsx
@@ -1,5 +1,4 @@
 import {
-    TitleBar,
     Typography,
 } from 'react-cm-ui';
 import {
@@ -23,8 +22,6 @@ function DocsActionBar() {
 
     return (
         <Main page={camelCase(displayName)}>
-            <TitleBar title={startCase(displayName)} />
-
             <Main.Content>
                 <MarkdownContainer>
                     <Typography