+++ docs/src/surfaces/card/card.jsx
@@ -1,8 +1,6 @@
-
-import PropTypes from 'prop-types';
 import React from 'react';
 import {
-    Card, Grid, Header, SubNavigation, TitleBar,
+    Card, Grid, Header,
 } from 'react-cm-ui';
 
 // Docs UI Components
@@ -211,10 +209,6 @@ export default class ViewsTitleBar extends React.Component {
     render() {
         return (
             <Main page="headers">
-                <TitleBar title="Card" />
-
-                <CardSubNavigation />
-
                 <Main.Content>
                     {this._renderCard()}
                 </Main.Content>