+++ docs/src/surfaces/filtersRail/filtersRail.jsx
@@ -1,30 +1,34 @@
 import {
-    TitleBar,
     Typography,
 } from 'react-cm-ui';
+import {
+    camelCase,
+} from 'lodash';
+import PropTypes from 'prop-types';
 import React from 'react';
 import ComponentApi from '../../global/componentApi';
 import Heading from '../../global/heading';
 import MarkdownContainer from '../../global/markdownContainer';
 import Main from '../../global/main';
 /* eslint-disable import/no-named-default, import/extensions */
-import { default as filtersRailDoc } from '!!@advclb/react-docgen-loader!react-cm-ui/surfaces/filtersRail/filtersRail';
+import { default as rootDoc } from '!!@advclb/react-docgen-loader!react-cm-ui/surfaces/filtersRail/filtersRail';
 /* eslint-enable import/no-named-default, import/extensions */
 
 function DocsFiltersRail() {
-    const descriptionCopy = filtersRailDoc.description;
+    const {
+        description,
+        displayName,
+    } = rootDoc;
 
     return (
-        <Main page="filters_rail">
-            <TitleBar title="Filters Rail" />
-
+        <Main page={camelCase(displayName)}>
             <Main.Content>
                 <MarkdownContainer>
                     <Typography
                         className="description"
                         variant="body1"
                     >
-                        {descriptionCopy}
+                        {description}
                     </Typography>
 
                     <Heading
@@ -37,7 +41,7 @@ function DocsFiltersRail() {
 
                 <ComponentApi
                     docs={[
-                        filtersRailDoc,
+                        rootDoc,
                     ]}
                 />
             </Main.Content>