+++ docs/src/utils/useMediaQuery/useMediaQuery.jsx
@@ -1,6 +1,3 @@
-import {
-    TitleBar,
-} from 'react-cm-ui';
 import React from 'react';
 import Heading from '../../global/heading';
 import MarkdownContainer from '../../global/markdownContainer';
@@ -9,8 +6,6 @@ import Main from '../../global/main';
 function DocsUseMediaQuery() {
     return (
         <Main page="use_media_query">
-            <TitleBar title="Use Media Query" />
-
             <Main.Content>
                 <MarkdownContainer>
                     <Heading